Updated user list query test to populate database with several users, then deleting them, keeping database clean.
