+++ seed/populate-database.ts
@@ -1,10 +1,6 @@
-import * as dotenv from 'dotenv';
-dotenv.config({ path: `${process.cwd()}/.test.env` });
-
 import { faker } from '@faker-js/faker';
 import { UserInput } from '../src/inputs/user-input';
 import { AppDataSource } from '../src/data-source';
-import { initApolloServer } from '../src/apollo-server';
 import { User } from '../src/entity/User';
 import { encryptPassword } from '../src/encryptPassword';
 
@@ -19,7 +15,7 @@ function createRandomUser(): UserInput {
   };
 }
 
-async function populateDatabase() {
+export async function populateDatabase() {
   const randomUsers: UserInput[] = [];
 
   for (let index = 0; index < numberOfNewUsers; index++) {
@@ -31,11 +27,3 @@ async function populateDatabase() {
 
   await AppDataSource.getRepository(User).save(randomUsers);
 }
-
-async function initServer() {
-  await AppDataSource.initialize();
-  await initApolloServer();
-  await populateDatabase();
-}
-
-initServer();
